refactor: api controller
