+++ frontend/src/services/api.ts
@@ -1,5 +1,4 @@
 import { FolderTree } from "../models/FileTree";
-import { NodeWithMetadata } from "../models/NodeWithMetadata";
 import { UploadedObjectMetadata } from "../models/UploadedObjectMetadata";
 import { getAuthSession } from "../utils/auth";
 import { uploadFileContent } from "../utils/file";
@@ -17,7 +16,7 @@ export const ApiService = {
       throw new Error("No session");
     }
 
-    const response = await fetch(`${API_BASE_URL}/upload-file`, {
+    const response = await fetch(`${API_BASE_URL}/objects/file`, {
       method: "POST",
       body: JSON.stringify({
         data: await uploadFileContent(file),
@@ -55,7 +54,7 @@ export const ApiService = {
       formData.append(fileId, file);
     });
 
-    const response = await fetch(`${API_BASE_URL}/upload-folder`, {
+    const response = await fetch(`${API_BASE_URL}/objects/folder`, {
       method: "POST",
       body: formData,
       headers: {
@@ -73,7 +72,7 @@ export const ApiService = {
   fetchUploadedObjectMetadata: async (
     cid: string
   ): Promise<UploadedObjectMetadata> => {
-    const response = await fetch(`${API_BASE_URL}/metadata/${cid}`);
+    const response = await fetch(`${API_BASE_URL}/objects/${cid}`);
 
     if (!response.ok) {
       throw new Error(`Network response was not ok: ${response.statusText}`);
@@ -81,17 +80,8 @@ export const ApiService = {
 
     return response.json();
   },
-  fetchData: async (cid: string): Promise<NodeWithMetadata> => {
-    const response = await fetch(`${API_BASE_URL}/retrieve/${cid}/node`);
-
-    if (!response.ok) {
-      throw new Error(`Network response was not ok: ${response.statusText}`);
-    }
-
-    return response.json() as Promise<NodeWithMetadata>;
-  },
   fetchDataURL: (cid: string): string => {
-    return `${API_BASE_URL}/retrieve/${cid}`;
+    return `${API_BASE_URL}/objects/${cid}/download`;
   },
   searchHeadCID: async (
     query: string,
@@ -103,7 +93,7 @@ export const ApiService = {
     }
 
     const response = await fetch(
-      `${API_BASE_URL}/metadata/search/${query}?scope=${scope}`,
+      `${API_BASE_URL}/objects/search?query=${query}&scope=${scope}`,
       {
         headers: {
           Authorization: `Bearer ${session?.accessToken}`,
@@ -120,7 +110,7 @@ export const ApiService = {
     }
 
     const response = await fetch(
-      `${API_BASE_URL}/metadata/roots?scope=${scope}`,
+      `${API_BASE_URL}/objects/roots?scope=${scope}`,
       {
         headers: {
           Authorization: `Bearer ${session?.accessToken}`,
